# Optimized FastRegexMatcher when the regex contains a case insensitive alternation made with concats too