We have a customer running queries with case insensitive regular expression matchers similar to the following one:
```
label=~"(?i:(foo1|foo2|bar)"
```

When we receive such queries, we see almost 40% of ingesters CPU spent running the regex:

![Screenshot 2023-02-28 at 17 52 08](https://user-images.githubusercontent.com/1701904/221922143-2c063405-ab58-4fcf-b27d-1eb64b7894fd.png)

These types of regex are expected to be accelerated by `FastRegexMatcher`, in particular by `stringMatcherFromRegexp()` so why isn't the fast path triggered?

The reason is that a regex like `(?i:(foo1|foo2|bar)` is unrolled by the engine as an alternate **and** a concat because of the common prefix between `foo1` and `foo2`. `stringMatcherFromRegexp()` can handle the scenario the regex is case insensitive and has an alternate, but cannot handle the scenario it's case insensitive and has a concat too. In this PR I've done some changes to support the case insensitive concat too.

To support it, I'm returning whether the matchers are case sensitive or not from `findSetMatches()`. `findSetMatches()` returns nil matches if the regex has a mix of case sensitivity.

### Benchmark

I've introduced a couple of new benchmark cases, to show the different before/after:

```
name                                                                                                                                  old time/op    new time/op    delta
FastRegexMatcher/foo-12                                                                                                                 7.42ns ± 2%    6.84ns ± 1%   -7.83%  (p=0.023 n=3+3)
FastRegexMatcher/^foo-12                                                                                                                8.87ns ±28%    6.90ns ± 2%     ~     (p=0.254 n=3+3)
FastRegexMatcher/(foo|bar)-12                                                                                                           8.93ns ± 3%    8.54ns ± 3%     ~     (p=0.132 n=3+3)
FastRegexMatcher/foo.*-12                                                                                                               29.0ns ± 2%    29.1ns ± 3%     ~     (p=0.950 n=3+3)
FastRegexMatcher/.*foo-12                                                                                                               32.0ns ± 3%    31.8ns ± 1%     ~     (p=0.659 n=3+3)
FastRegexMatcher/^.*foo$-12                                                                                                             31.5ns ± 2%    32.2ns ± 1%     ~     (p=0.151 n=3+3)
FastRegexMatcher/^.+foo$-12                                                                                                             31.6ns ± 2%    32.1ns ± 2%     ~     (p=0.321 n=3+3)
FastRegexMatcher/.*-12                                                                                                                  26.3ns ± 0%    26.2ns ± 1%     ~     (p=0.585 n=3+3)
FastRegexMatcher/.+-12                                                                                                                  27.2ns ± 1%    26.1ns ± 3%     ~     (p=0.087 n=3+3)
FastRegexMatcher/foo.+-12                                                                                                               30.3ns ± 4%    28.1ns ± 2%   -7.43%  (p=0.049 n=3+3)
FastRegexMatcher/.+foo-12                                                                                                               31.7ns ± 1%    32.7ns ± 1%   +3.19%  (p=0.014 n=3+3)
FastRegexMatcher/.*foo.*-12                                                                                                             79.4ns ± 4%    79.8ns ± 1%     ~     (p=0.830 n=3+3)
FastRegexMatcher/(?i:foo)-12                                                                                                            21.4ns ± 1%    21.0ns ± 1%     ~     (p=0.187 n=3+3)
FastRegexMatcher/(?i:(foo|bar))-12                                                                                                      42.1ns ± 2%    41.7ns ± 3%     ~     (p=0.684 n=3+3)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12                                                                                                 168ns ± 2%      66ns ± 1%  -60.73%  (p=0.000 n=3+3)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|eee|fff|ggg|hhh|iii|lll|mmm|nnn|ooo|ppp|qqq|rrr|sss|ttt|uuu|vvv|zzz|xxx|yyy|www))-12    1.57µs ± 3%    0.39µs ± 2%  -74.98%  (p=0.000 n=3+3)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-12                                                                                     45.8ns ± 1%    45.0ns ± 1%     ~     (p=0.086 n=3+3)
FastRegexMatcher/((fo(bar))|.+foo)-12                                                                                                   58.4ns ± 1%    60.0ns ± 1%   +2.67%  (p=0.019 n=3+3)

name                                                                                                                                  old alloc/op   new alloc/op   delta
FastRegexMatcher/foo-12                                                                                                                  0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/^foo-12                                                                                                                 0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(foo|bar)-12                                                                                                            0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/foo.*-12                                                                                                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.*foo-12                                                                                                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/^.*foo$-12                                                                                                              0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/^.+foo$-12                                                                                                              0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.*-12                                                                                                                   0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.+-12                                                                                                                   0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/foo.+-12                                                                                                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.+foo-12                                                                                                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.*foo.*-12                                                                                                              0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:foo)-12                                                                                                             0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:(foo|bar))-12                                                                                                       0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12                                                                                                 0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|eee|fff|ggg|hhh|iii|lll|mmm|nnn|ooo|ppp|qqq|rrr|sss|ttt|uuu|vvv|zzz|xxx|yyy|www))-12     0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-12                                                                                      0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/((fo(bar))|.+foo)-12                                                                                                    0.00B          0.00B          ~     (zero variance)

name                                                                                                                                  old allocs/op  new allocs/op  delta
FastRegexMatcher/foo-12                                                                                                                   0.00           0.00          ~     (zero variance)
FastRegexMatcher/^foo-12                                                                                                                  0.00           0.00          ~     (zero variance)
FastRegexMatcher/(foo|bar)-12                                                                                                             0.00           0.00          ~     (zero variance)
FastRegexMatcher/foo.*-12                                                                                                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/.*foo-12                                                                                                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/^.*foo$-12                                                                                                               0.00           0.00          ~     (zero variance)
FastRegexMatcher/^.+foo$-12                                                                                                               0.00           0.00          ~     (zero variance)
FastRegexMatcher/.*-12                                                                                                                    0.00           0.00          ~     (zero variance)
FastRegexMatcher/.+-12                                                                                                                    0.00           0.00          ~     (zero variance)
FastRegexMatcher/foo.+-12                                                                                                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/.+foo-12                                                                                                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/.*foo.*-12                                                                                                               0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:foo)-12                                                                                                              0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:(foo|bar))-12                                                                                                        0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12                                                                                                  0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|eee|fff|ggg|hhh|iii|lll|mmm|nnn|ooo|ppp|qqq|rrr|sss|ttt|uuu|vvv|zzz|xxx|yyy|www))-12      0.00           0.00          ~     (zero variance)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-12                                                                                       0.00           0.00          ~     (zero variance)
FastRegexMatcher/((fo(bar))|.+foo)-12                                                                                                     0.00           0.00          ~     (zero variance)
```