@pracucci 2023-03-01
Going to merge this PR and then will work on fuzzy testing in a follow up PR. I've already found an issue with a basic fuzzy test, but the issue happens in `main` too so it's unrelated from this PR.